// app/types/experience.type.ts
export type IExperience = {
  id: string;
  name: string;
  size: string;
  type: string;
  workType: string;
  designation: string;
  workLocation: string;
  location: string;
  startDate: string;
  endDate: string;
  website: string;
  linkedIn: string;
  responsibilities: string[];
  learnedNewTech: string[];
};